Am legat, plateRecognition algoritmul vechi cu parserul de numar si comunicarea seriala prin USART cu arduino
